Create core game logic for playing one round
Program converts both answers to lowercase and then compares them. If
they match, it declares a tie and no score change takes place. Otherwise
, the human answer is fed into a switch-case statement and the win
condition variable assigned true or false with a conditional statement.
I created a another function that accepts the win condition as an
argument and then manages the scores.